requirements.txt
sentence-transformers==5.3.0
docling==2.80.0
JSON-log-formatter==1.1.1
cohere==5.20.7
python-dotenv==1.2.2
pymupdf==1.27.2.2
